requirements.txt
streamlit>=1.22.0
markdown>=3.4.3
Pillow>=9.4.0
python-markdown-math>=0.8
requests>=2.28.2
nltk>=3.8.1
networkx>=3.0
scipy>=1.10.1
PyMuPDF>=1.21.1
